+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [1.14.2] - 2024-02-27
 ### Added
 - Short usage and help synopsis using the short help option (`-h`).
 - Application configuration aliases (`alias:`) as a solution to tricky
@@ -564,7 +565,8 @@ Major release.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/util/compare/v1.14.1...HEAD
+[Unreleased]: https://github.com/plandes/util/compare/v1.14.2...HEAD
+[1.14.2]: https://github.com/plandes/util/compare/v1.14.1...v1.14.2
 [1.14.1]: https://github.com/plandes/util/compare/v1.14.0...v1.14.1
 [1.14.0]: https://github.com/plandes/util/compare/v1.13.3...v1.14.0
 [1.13.3]: https://github.com/plandes/util/compare/v1.13.2...v1.13.3